Made changes im dir
